apps/api-gateway/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# AI/LLM
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.20
llama-index==0.9.48

# Vector Database & Embeddings
qdrant-client==1.7.3
sentence-transformers==2.3.1
torch==2.1.2
numpy==1.26.3

# Task Queue
celery==5.3.6
flower==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pycryptodome==3.19.1

# Monitoring
prometheus-client==0.19.0

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
ruff==0.1.14
black==24.1.1
mypy==1.8.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
